remove old functions, separate constants
