+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp_Main.java
@@ -12,14 +12,6 @@ public class TeleOp_Main extends Base {
 
         while (active()) {
             drivetrainLogic(p != null, false);
-            wristMotorLogic();
-            wristServoXLogic();
-            intakeLogic();
-            horizontalLiftLogic();
-            verticalLiftLogic();
-            basketServoLogic();
-            specimenServoLogic();
-            handoffLogic();
             dpadLogic();
 
             telemetryAll();
